Argument testing and os independency fixes

+++ mxcppfy.py
@@ -8,6 +8,21 @@ def merge_files(filename_, directory_='.'):
     cpp_bak_file = os.path.join(directory_, filename_ + '.cpp.bak')
     c_file = os.path.join(directory_, filename_ + '.c')
 
+    # Checking that at least the c file exists
+    if not os.path.exists(c_file):
+        print(f'{c_file} cannot be found.')
+        exit(1)
+
+    # If no cpp file exists, the file is simply copied from the c file
+    if not os.path.exists(cpp_file):
+        os.replace(c_file, cpp_file)
+        print(f'{cpp_file} has been created from {c_file}!')
+        exit(0)
+
+    # Checking if a previous .bak exists and removing it
+    if os.path.exists(cpp_bak_file):
+        os.remove(cpp_bak_file)
+
     # Rename .cpp to .cpp.bak
     os.rename(cpp_file, cpp_bak_file)
 
@@ -32,10 +47,9 @@ def merge_files(filename_, directory_='.'):
                         if f'/* USER CODE END {block_name}' in cpp_bak_line:
                             break
                 # We can skip lines until we evenntually get to the right block name
-                pass
             elif copying_cpp:
                 # If we reach the end of a USER CODE section we can resume copying from the .c file
-                if f'/* USER CODE END' in line:
+                if '/* USER CODE END' in line:
                     copying_cpp = False
             else:
                 # Standard copy from .c
@@ -49,7 +63,7 @@ if len(sys.argv) < 2:
 
 filename = sys.argv[1]
 # Defaults to current wd
-directory = sys.argv[2] if len(sys.argv) > 2 else '.'
+directory = sys.argv[2] if len(sys.argv) > 2 else f'.{os.sep}Core{os.sep}Src'
 
 merge_files(filename, directory)
